Bunch of fixes to the listview code to make it work with the real e-mission app
The most important one of these is the one where I remove ng-app="starter" from
the body - mainly because I had already removed this once before, and it was
added back when you were adding changes. It is really important to have that
working, at least on android, because we can only initialize the sqlite plugin
after the document is ready, and we can only load the data after the sqlite
plugin has been initialized. So you really need to load the starter only after
the document is ready. I have added a separate script in the listview to do this.

Other fixes:
- Change the location of the database on iOS to the same one as the native code
- Move the android tabs to the bottom so that they don't cover the date
- Remove spaces from the base64 encoding of the trip JSON so that it can be
  properly decoded by android
- Display error information if there are errors while decoding