Support both android and ios native formats
It looks like android and ios both store and retrieve the data slightly
differently. On iOS, we need to base64 decode the returned value, but on
android, we don't. In order to ensure that we test with both sets of data, we:
- delete the main TripSections.db
- replace it with android and iOS specific TripSections.db

The tests should really be run with both.

Also change the code to handle both cases in a fairly naive way (try/catch and
use the alternate approach) and ensure that the test cases pass.