Deal with the fallout of the upgrade to the latest version of ionic/cordova
It turns out that the most recent version of ionic/cordova uses
cordova-android@4.0.0. This requires the use of the whitelist plugin to enable
access to non-file URLs.
http://cordova.apache.org/announcements/2015/04/15/cordova-android-4.0.0.html

Even after the plugin is installed, it only supports access to file:// URLs.
We need to explicitly open access to other sites in the config.xml

Doing this in the most naive way right now to unblock Jillie.
We need to think about what our tradeoff between flexibility and functionality
should be - with this, calls to other sites will require an app upgrade.

Maybe a proxying service through our own server would be a good compromise.