Remove the code to create the slidebox
While it would be great to pre-load data, creating all those maps at the same
time kills (at least) my phone and visibly affects the performance of Tom's.
In particular, because of the low memory, the map tiles take a long time to
load, and the z-index is messed up, so that they are displayed on top of the
geojson layer.

Change this to remove the slidebox, implement prev and next buttons manually
and lazily load the data. Once we have the server side call to retrieve more
data on the fly, we can hook up dynamic retrieval as well.