Add the ability to create a magnifying glass on the map by pressing and holding
I really think that the magnifying glass is a good way to get clarity on parts
of the map that are hiding a lot of detail. However, I haven't been able to
figure out how to do it properly. Having a fixed glass at the end points is not
flexible enough, and also seems to have issues with duplication.

With this, I try to create a glass which can be moved to the desired location.
On playing with it, it is still a bit hard to use because of the offset, and
the extreme sensitivity of the touch screen.

Leaving it in so that people can play with it.

If we choose to retain it, the relevant changes need to be moved out, at least
of angular-leaflet-directive. I should also consider submitting a pull request
to magnifying glass to add support for mobile touch devices.