Switch the checked in sample data to the new format
It is now separated by date, to make it easier to retrieve data one day at a
time. This allows us to improve both the memory profile and the load time - we
no longer need to bin trips on the client, and we can also load data one day at
a time and allow data from prior days to be garbage collected.