UI for the tracking part
In a real application, this would be a debug screen. The intent is to show
values that have been collected but not yet transmitted, and the operation of
the state machine.

Note that this can be considerably more complex (animation of state
transitions?!) but is probably good enough as a debugging tool for now.