Convert the native screens into their own little plugin
This ensures that they are copied properly without adding a separate hook. More
importantly, on iOS, it ensures that the files are added to the project, and
are found during compilation. Without that, the .xib file is not added to the
project, and the screen is not displayed.

As an aside, it looks like it works for a plugin to have no javascript component.